Typescript structure and redux types
