--- Client/src/Redux/actions/types.js
@@ -1,19 +1,0 @@
-export const GET_ALL_PRODUCTS="GET_ALL_PRODUCTS";
-export const GET_ALL_PRODUCTS_ERROR="GET_ALL_PRODUCTS_ERROR";
-export const GET_ALL_FAVS="GET_ALL_FAVS";
-export const GET_ALL_FAVS_ERROR="GET_ALL_FAVS_ERROR";
-
-export const ADD_TO_FAVS = "ADD_TO_FAVS";
-export const ADD_TO_FAVS_ERROR = "ADD_TO_FAVS_ERROR";
-export const OUT_OF_FAVS="OUT_OF_FAVS";
-
-export const ADD_TO_CART = "ADD_TO_CART";
-export const REMOVE_FROM_CART = "REMOVE_FROM_CART";
-
-export const ADJUST_MORE_QTY = " ADJUST_MORE_QTY";
-export const ADJUST_LESS_QTY = "ADJUST_LESS_QTY";
-export const ADJUST_ITEM_QTY = "ADJUST_ITEM_QTY";
-
-export const PAYDAY = "PAYDAY";
-export const PAYDAY_ERROR = "PAYDAY_ERROR";
-
