--- Client/src/api/products.js
@@ -1,6 +1,0 @@
-// import types
-import axios from "axios";
-
-export default axios.create({
-  baseURL: "http://localhost:3000/",
-});
