--- Client/src/components/Cart/Cart.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Cart = () => {
-    return (
-        <div>
-            cart
-        </div>
-    )
-}
-
-export default Cart
