--- Client/src/components/Favs/Favs.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Favs = () => {
-    return (
-        <div>
-            favs
-        </div>
-    )
-}
-
-export default Favs
