--- Client/src/components/Products/Product/Product.scss
@@ -1,43 +1,0 @@
-@import '../../../scss/index.scss';
-
-.product-container {
-    color:$defaultDarkColorFont;
-    background-color: $defaultLightColorFont;
-    
-    height: auto;
-    width: 350px;
-
-    padding: 15px;
-    margin: 5px;
-
-    display: flex;
-    flex-direction: column;
-
-    border-radius: 5%;
-
-    &__img {
-        img {
-            min-width: 280px;
-            height: 280px;
-            border-radius: 3%;
-            margin-bottom: 20px;
-        }
-        // min-width: 280px;
-        // height: 200px;
-        // border-radius: 3%;
-        // margin-bottom: 20px;
-
-    }
-
-    &__info {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-evenly;
-    }
-    &__add {
-        img {
-            width: 30px;
-            height: 30px;;
-        }
-    }
-}
