+++ client/README.md
@@ -1,0 +1,11 @@
+# Getting Started with Create React App
+```
+npx create-react-app client --template typescript
+
+
+ yarn add redux-thunk redux react-redux redux-devtools-extension bootstrap react-router-dom reactstrap sass axios
+
+ yarn add -D @types/redux @types/react-redux @types/redux-thunk
+
+
+```
