+++ client/src/Redux/actions/types.ts
@@ -1,0 +1,21 @@
+type action = string;
+
+export const GET_ALL_PRODUCTS: action="GET_ALL_PRODUCTS";
+export const GET_ALL_PRODUCTS_ERROR: action="GET_ALL_PRODUCTS_ERROR";
+export const GET_ALL_FAVS: action="GET_ALL_FAVS";
+export const GET_ALL_FAVS_ERROR: action="GET_ALL_FAVS_ERROR";
+
+export const ADD_TO_FAVS: action= "ADD_TO_FAVS";
+export const ADD_TO_FAVS_ERROR: action= "ADD_TO_FAVS_ERROR";
+export const OUT_OF_FAVS: action="OUT_OF_FAVS";
+
+export const ADD_TO_CART: action= "ADD_TO_CART";
+export const REMOVE_FROM_CART: action= "REMOVE_FROM_CART";
+
+export const ADJUST_MORE_QTY: action= " ADJUST_MORE_QTY";
+export const ADJUST_LESS_QTY: action= "ADJUST_LESS_QTY";
+export const ADJUST_ITEM_QTY: action= "ADJUST_ITEM_QTY";
+
+export const PAYDAY: action= "PAYDAY";
+export const PAYDAY_ERROR: action= "PAYDAY_ERROR";
+
