+++ client/src/api/products.ts
@@ -1,0 +1,6 @@
+// import types
+import axios from "axios";
+
+export default axios.create({
+  baseURL: "http://localhost:5000/",
+});
