+++ client/src/components/Navbar/Pages/MainPage/MainPage.scss
@@ -1,0 +1,26 @@
+@import '../../scss/index.scss';
+
+.main-page {
+    background-color: $backgroundColor;
+    height: 100%;
+    width: 100%;
+    position: absolute;
+    
+
+    top: 0;
+    left: 0;
+    
+    text-align: center;
+    color:$defaultLightColorFont;
+
+    &__title {
+        margin-top: 50px;
+        font-size:$bigFontSize;
+    }
+    &__subtitle {
+        font-size:$smallFontSize;
+    }
+
+
+    
+}
